lambdas/requirements.txt
requests
python-jose
boto3